Made two more objects of the class Face. Updated class Face with points param. Updated TODO
